Move federationin api content
