+++ cmd/monolith/main.go
@@ -21,12 +21,12 @@ import (
 	"net/http"
 
 	"github.com/google/exposure-notifications-server/internal/api/export"
-	"github.com/google/exposure-notifications-server/internal/api/federationin"
 	"github.com/google/exposure-notifications-server/internal/api/handlers"
 	"github.com/google/exposure-notifications-server/internal/api/publish"
 	cleanup "github.com/google/exposure-notifications-server/internal/cleanup/api"
 	"github.com/google/exposure-notifications-server/internal/database"
 	"github.com/google/exposure-notifications-server/internal/dbapiconfig"
+	"github.com/google/exposure-notifications-server/internal/federationin"
 	"github.com/google/exposure-notifications-server/internal/logging"
 	"github.com/google/exposure-notifications-server/internal/setup"
 )
